[Mac] Remove NSURLCache logic from WebCore pressure relief code. <https://webkit.org/b/130248>
The NSURLCache lives in the networking process and should be cleared
on that end. Furthermore, this code doesn't really do what we thought
it was doing; it merely sets and and resets the limits. No truncation
ever occurred. The intended functionality was implemented in r165342.

Reviewed by Anders Carlsson.

* platform/mac/MemoryPressureHandlerMac.mm:
(WebCore::MemoryPressureHandler::releaseMemory):

Identifier: 148220@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165631 268f45cc-cd09-0410-ab3c-d52691b4dbfc