[CSSRegions]Do not compute region range for a box unless the parent has one https://bugs.webkit.org/show_bug.cgi?id=130249
Reviewed by Andrei Bucur.

Source/WebCore:

If the containing block does not have a region range computed, do not attempt to compute
a region range for a child. In such cases, the range computation for a child can lead to
a result that is wrong, possibly leading to an incorrect clipping.

Test: fast/regions/inline-block-flowed-in-regions.html

* rendering/RenderBlock.cpp:
(WebCore::canComputeRegionRangeForBox):
(WebCore::RenderBlock::computeRegionRangeForBoxChild):
(WebCore::RenderBlock::estimateRegionRangeForBoxChild):
(WebCore::RenderBlock::updateRegionRangeForBoxChild):
* rendering/RenderFlowThread.h:

LayoutTests:

* fast/regions/inline-block-flowed-in-regions-expected.html: Added.
* fast/regions/inline-block-flowed-in-regions.html: Added.


Identifier: 148303@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165720 268f45cc-cd09-0410-ab3c-d52691b4dbfc