[iOS] WebPDFViewPlaceholder divides-by-zero with an empty document https://bugs.webkit.org/show_bug.cgi?id=130444 <rdar://problem/15580682>
Reviewed by Sam Weinig.

* WebView/WebPDFViewPlaceholder.h:
* WebView/WebPDFViewPlaceholder.mm:
(-[WebPDFViewPlaceholder layout]):
Get rid of boundingSize, it's never assigned to anything but CGSizeZero,
which is confusing.

Don't divide by zero if we have no pages, instead return an empty bounding size.

(-[WebPDFViewPlaceholder _computePageRects:]):
If we have an empty document, set the background color to white.
Otherwise, set it to black. It's not clear that black is the ideal
background color, but this retains existing behavior.

* UIProcess/ios/WKPDFView.mm:
(-[WKPDFView web_setContentProviderData:]):
Don't make UIPDFPageViews for null pages.


Identifier: 148447@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@165896 268f45cc-cd09-0410-ab3c-d52691b4dbfc