6% SunSpider commandline regression due to r165940 https://bugs.webkit.org/show_bug.cgi?id=130617
Reviewed by Michael Saboff.

In GCActivityCallback::didAllocate, lastGCLength() returns 0 if we've never collected
before. Some of the benchmarks are never running a single EdenCollection, which causes
them to repeatedly call scheduleTimer with a newDelay of 0. This defeats our timer
slop heuristic, causing us to invoke CFRunLoopTimerSetNextFireDate a couple orders of
magnitude more than we normally would.

The fix is to seed the last GC lengths in Heap with a non-zero length so that our heuristic works.

* heap/Heap.cpp:
(JSC::Heap::Heap):


Identifier: 148640@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166099 268f45cc-cd09-0410-ab3c-d52691b4dbfc