-[WKWebProcessPlugInNodeHandle HTMLInputElementIsUserEdited] and -HTMLTextAreaElementIsUserEdited do not return the correct values.
https://bugs.webkit.org/show_bug.cgi?id=130804

Patch by Cody Krieger <cjk@apple.com> on 2014-03-26
Reviewed by Dan Bernstein.

The existing implementations of these methods are mis-cased
("HTMLInputELementIsUserEdited" and "HTMLTextAreaELementIsUserEdited" —
note the "EL" in both cases). This results in us getting
auto-synthesized getters for the HTMLInputElementIsUserEdited and
HTMLTextAreaElementIsUserEdited properties, which will return NO
unconditionally.

* WebProcess/InjectedBundle/API/Cocoa/WKWebProcessPlugInNodeHandle.mm:
Fix the capitalization of -HTMLInputElementIsUserEdited and
-HTMLTextAreaElementIsUserEdited.

Identifier: 148843@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166326 268f45cc-cd09-0410-ab3c-d52691b4dbfc