[CSSRegions] Crash when cloning a region child with a content node child https://bugs.webkit.org/show_bug.cgi?id=129811
Reviewed by David Hyatt.

Source/WebCore:

Collecting the children of a region in another named flow will be supported
again in the future and it will be revisited when we will implement the content
keyword: http://dev.w3.org/csswg/css-regions-1/#the-flow-into-property.
Remove the support for now as it introduces unnecessary complexity in the code
and potential wrong behavior.

Test: fast/regions/region-content-node-child-clone-crash.html

* dom/Element.cpp:
(WebCore::Element::shouldMoveToFlowThread):
(WebCore::Element::clearStyleDerivedDataBeforeDetachingRenderer):
* dom/Element.h:
* dom/ElementRareData.h:
(WebCore::ElementRareData::ElementRareData):
* rendering/FlowThreadController.cpp:
* rendering/FlowThreadController.h:
* style/StyleResolveTree.cpp:
(WebCore::Style::moveToFlowThreadIfNeeded):
(WebCore::Style::createRendererIfNeeded):

LayoutTests:

Collecting the children of a region in another named flow will be supported
again in the future and it will be revisited when we will implement the content
keyword: http://dev.w3.org/csswg/css-regions-1/#the-flow-into-property.
For now, skip the tests for that use case.

* TestExpectations:
* fast/regions/overflow-region-inline-expected.html: Expected file should not be region based.
* fast/regions/region-content-node-child-clone-crash-expected.txt: Added.
* fast/regions/region-content-node-child-clone-crash.html: Added.
* fast/regions/unsplittable-nested-region-expected.html: Expected file should not be region based.


Identifier: 148869@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166353 268f45cc-cd09-0410-ab3c-d52691b4dbfc