[GTK] Running minibrowser with url crashes in debug build https://bugs.webkit.org/show_bug.cgi?id=130492
Reviewed by Martin Robinson.

Use the ResourceRequest(const URL& url) constructor, instead of the
one taking a String, to call WebPageProxy::loadRequest(), so we
ensure that the passed URI is correctly constructed.

* UIProcess/API/gtk/WebKitWebView.cpp:
(webkit_web_view_load_uri): Create an URL out of a SoupURI and
use that to create a ResourceRequest to pass to loadRequest().


Identifier: 148926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166410 268f45cc-cd09-0410-ab3c-d52691b4dbfc