Improve GC_LOGGING https://bugs.webkit.org/show_bug.cgi?id=130988
Reviewed by Geoffrey Garen.

GC_LOGGING can be useful for diagnosing where we're spending our time during collection,
but it doesn't distinguish between Eden and Full collections in the data it gathers. This
patch updates it so that it can. It also adds the process ID to the beginning of each line
of input to be able to distinguish between the output of multiple processes exiting at the
same time.

* heap/Heap.cpp:
(JSC::Heap::collect):


Identifier: 149054@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166544 268f45cc-cd09-0410-ab3c-d52691b4dbfc