ASSERT(m_children.isEmpty()) in PlatformCALayerRemote::removeAllSublayers() on time.com https://bugs.webkit.org/show_bug.cgi?id=131187
Reviewed by Maciej Stachowiak.

After a call to PlatformCALayerRemote::adoptSublayers() when a layer
toggles between tiled and non-tiled, we'd assert in PlatformCALayerRemote::removeAllSublayers()
that not all sublayers had been removed (there was one with a null superlayer).

Fix by ensuring that we don't copy over custom subviews when adoptSublayers()
is called; they'll get added as we update children anyway.

* WebProcess/WebPage/mac/PlatformCALayerRemote.cpp:
(WebKit::PlatformCALayerRemote::adoptSublayers):

Identifier: 149251@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166753 268f45cc-cd09-0410-ab3c-d52691b4dbfc